+++ packages/frontend/src/config/index.ts
@@ -1,5 +1,7 @@
 export const Port = 3000;
 
-export const CakeApiUrl = 'http://localhost:3100/cakes';
+export const ApiHost = 'http://localhost:3100';
+
+export const CakeApiUrl = `${ApiHost}/cakes`;
 
 export const CakeFormName = 'CakeForm';